feat: return uint64 from XDM's baseGas
ctb: Add a test showing max message length for baseGas

+++ packages/contracts-bedrock/contracts/L1/L1CrossDomainMessenger.sol
@@ -44,7 +44,7 @@ contract L1CrossDomainMessenger is CrossDomainMessenger, Semver {
      */
     function _sendMessage(
         address _to,
-        uint64 _gasLimit,
+        uint32 _gasLimit,
         uint256 _value,
         bytes memory _data
     ) internal override {
